+++ components/Footer/Contact.js
@@ -1,0 +1,10 @@
+import style from "./style.scss";
+
+const Contact = () => (
+  <div className={style.contact}>
+    <h2>Contact Us</h2>
+    <h1>contact@warsawex.org</h1>
+  </div>
+);
+
+export default Contact;
