+++ components/Founders/index.js
@@ -5,12 +5,12 @@ const Founders = () => (
     <h1>Our Founders</h1>
     <hr />
     <div className={style.thesegoys}>
-      <div className={style.kamil}>
+      <div className={style.founder}>
         <img src="/static/kamil_kowalski.png" alt="photo of Kamil Kowalski" />
         <h2>Kamil Kowalski</h2>
         <h3>Engineering Manager at Shedul</h3>
       </div>
-      <div className={style.robert}>
+      <div className={style.founder}>
         <img
           src="/static/kamil_kowalski.png"
           alt="photo of Robert Maciejewski"
