+++ components/Header/Logo.js
@@ -1,0 +1,11 @@
+import Link from "next/link";
+
+const Logo = () => (
+  <Link href="/">
+    <a>
+      <img src="/static/warsawex-logo.svg" alt="warsaw.ex logo" />
+    </a>
+  </Link>
+);
+
+export default Logo;
