+++ components/Header/style.scss
@@ -46,6 +46,9 @@
             font-size: 20px;
             font-weight: 500;
             margin: 0px 15px 0px 15px;
+            a:hover {
+              cursor: pointer;
+            }
           }
         }
         > .media {
